Disabled the -i argument in Scala 3 modules
